+++ video_noise/scene_interference.py
@@ -146,7 +146,7 @@ def add_rain_fractal(video: torch.Tensor, ratio: float, severity: int = 5) -> to
             frame_tensor = torch.from_numpy(rainy_frame).permute(2, 0, 1)
         
         # 更新原视频
-        video[i] = frame_tensor.to(device).to(video.dtype)
+        video[i] = frame_tensor.to(device).to(torch.uint8)
     
     return video
 # def add_rain(video: torch.Tensor, ratio: float) -> torch.Tensor:
@@ -159,7 +159,7 @@ def add_rain_fractal(video: torch.Tensor, ratio: float, severity: int = 5) -> to
 
 #     for i in noise_indices:
 #         synthetic = alpha * rain_layer + belta * video[i]
-#         video[i] = torch.clamp(synthetic, 0, 255).to(video.dtype)
+#         video[i] = torch.clamp(synthetic, 0, 255).to(torch.uint8)
 #     return video
 
 
@@ -266,7 +266,7 @@ def add_fog_fractal(video: torch.Tensor, ratio: float, severity: int = 5) -> tor
             frame_tensor = torch.from_numpy(frame_np).permute(2, 0, 1)
         
         # 更新原视频
-        video[i] = frame_tensor.to(device).to(video.dtype)
+        video[i] = frame_tensor.to(device).to(torch.uint8)
     
     return video
 # def add_rain(video: torch.Tensor, ratio: float) -> torch.Tensor:
@@ -279,7 +279,7 @@ def add_fog_fractal(video: torch.Tensor, ratio: float, severity: int = 5) -> tor
 
 #     for i in noise_indices:
 #         synthetic = alpha * rain_layer + belta * video[i]
-#         video[i] = torch.clamp(synthetic, 0, 255).to(video.dtype)
+#         video[i] = torch.clamp(synthetic, 0, 255).to(torch.uint8)
 #     return video
 
 @NoiseRegistry.register("snow") # 雪
@@ -442,7 +442,7 @@ def add_snow_effect(video: torch.Tensor, ratio: float, severity: int = 5) -> tor
             frame_tensor = torch.from_numpy(snowy_frame).permute(2, 0, 1)
         
         # 更新原视频
-        video[i] = frame_tensor.to(device).to(video.dtype)
+        video[i] = frame_tensor.to(device).to(torch.uint8)
     
     return video
 
@@ -608,7 +608,7 @@ def add_frost_effect(video: torch.Tensor, ratio: float, severity: int = 5, frost
             frame_tensor = torch.from_numpy(frosted_frame).permute(2, 0, 1)
         
         # 更新原视频
-        video[i] = frame_tensor.to(device).to(video.dtype)
+        video[i] = frame_tensor.to(device).to(torch.uint8)
     
     return video
 
@@ -645,7 +645,7 @@ def add_specular_reflection(video: torch.Tensor, ratio: float) -> torch.Tensor:
         frame = 0.9 * frame + 0.1 * noise_layer
         frame = torch.clamp(frame, 0, 255) 
         
-        video[i] = frame.to(video.dtype) 
+        video[i] = frame.to(torch.uint8) 
 
     return video
